backend/requirements.txt
# Core API dependencies
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9

# Database dependencies
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0
greenlet==3.0.3

# Data processing
pandas==2.2.0
numpy==1.26.3

# Security dependencies
cryptography==42.0.8
pyjwt==2.10.1
passlib==1.7.4
python-jose==3.3.0

# Validation
email-validator==2.2.0
tzdata==2024.2

# HTTP clients
requests==2.31.0
httpx==0.28.0


# Development and testing
pytest==8.2.0
pytest-asyncio==1.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# MQTT integration
paho-mqtt==1.6.1

# MongoDB support for database testing
pymongo==4.7.1

# Prometheus metrics
prometheus-client==0.16.0

# Load testing
locust==2.24.0

# Optional dependencies (remove if not using)
# boto3==1.34.129  # For AWS services
# requests-oauthlib==2.0.0  # For OAuth flows
